requirements.txt
mysql-connector-python>=9.0
requests>=2.28
openpyxl>=3.1
matplotlib>=3.7
